90% - No token middleware bug (#18)
* Fix middleware from throwing exception

ValidateToken now throws an exception if `token` is not supplied.
The token is checked by using `_.isEmpty` (null, "", 0, etc). As the
token is supplied by the user within the middleware it is possible that
it could be null or a empty string. This will throw an exception that
is not handled correctly.

* Provide name to missing config exceptions

* Hide internal exception message

An internal exception message could expose internal workings
to an external hacker if exposed through a http error response.