Return decoded token upon successful validation
This change returns a decoded JWT upon successful validation instead of the string "ok".

This is so that metadata can be extracted and used by other services. For example, when securing an AWS API gateway route, Amazon needs to know who the token belongs to in order to create an authentication policy for the request.