Return decoded token upon un/successful validation (#21)
This change returns a decoded JWT upon un/successful validation as a third argument to the returned callback.

This change supersedes #20 by making this a non-breaking change.

This is so that metadata can be extracted and used by other services. For example, when securing an AWS API gateway route, Amazon needs to know who the token belongs to in order to create an authentication policy for the request.