Remove test
We don't use the secret as part of the cache key as it doesn't make
sense. Including the secret in the key doesn't increase security and it
could expose more information than we would want. As we cannot afford to
loose the secret, we should not use it as a cache key.