PLAT-963 - Unique cache keys (#29)
* postfix the cache keys with the persona url

* syntax fixes

* Remove test

We don't use the secret as part of the cache key as it doesn't make
sense. Including the secret in the key doesn't increase security and it
could expose more information than we would want. As we cannot afford to
loose the secret, we should not use it as a cache key.

* Use correct cache key in test

* change cacheGetKey -> _formatCacheKey

* DRY the code

* bump version to 3.4.0

As the cache key has changed I have bounced the version from 3.3.1 to
3.4.0. This will invalidate all cached items.

* lint fixes

* remove wrong method name

* test key format is different per config

* Check key format